Enable some tests in test_module_payload_dnf5_manager
These tests probably didn't work at the time they were marked as
skipped, but are working now.
